+++ packages/react-components-core/src/config/createConfig.types.ts
@@ -1,13 +1,13 @@
 /* eslint-disable @typescript-eslint/no-empty-interface */
 import { Config } from "./config.types";
-import { RecursivePartial } from "@peersyst/react-types";
+import { DeepPartial } from "@peersyst/react-types";
 
 export interface BaseCoreCreateConfig {
     projectName: string;
     themes?: Partial<Config["themes"]>;
     translate?: Config["translate"];
     validators?: Config["validators"];
-    components?: RecursivePartial<Config["components"]>;
+    components?: DeepPartial<Config["components"]>;
 }
 
 export interface CoreCreateConfig extends BaseCoreCreateConfig {}