+++ packages/react-components-core/src/theme/createTheme.types.ts
@@ -8,6 +8,7 @@ import {
     TypographyType,
     UnitsType,
 } from "./theme.types";
+import { DeepPartial } from "@peersyst/react-types";
 
 export type CreateThemePalette = Partial<ThemePalette>;
 
@@ -19,5 +20,8 @@ export interface CreateCoreTheme {
     borderRadius: UnitsType;
     zIndex?: Partial<ThemeZIndex>;
 }
-export interface CoreCreateTheme extends Partial<CreateCoreTheme> {}
+export interface CoreCreateTheme extends Omit<Partial<CreateCoreTheme>, "typography" | "icons"> {
+    icons?: Partial<CreateCoreTheme["icons"]>;
+    typography?: DeepPartial<CreateCoreTheme["typography"]>;
+}
 export interface CreateTheme extends Partial<CoreCreateTheme> {}