+++ packages/react-native-components/src/theme/index.ts
@@ -7,4 +7,17 @@ export * from "./shadows";
 export * from "./typography";
 export { default as ThemeProvider } from "./ThemeProvider";
 export { useTheme, ThemeContext, ThemeConsumer } from "@peersyst/react-components-core";
-export type { ThemeContextType, Theme } from "@peersyst/react-components-core";
+export type {
+    ThemeContextType,
+    Theme,
+    CreateTheme,
+    ThemePalette,
+    ThemeFonts,
+    ThemeZIndex,
+    TypographyVariants,
+    TypographyVariantsOverrides,
+    ThemeTypography,
+    ThemeIcons,
+    PaletteMode,
+} from "@peersyst/react-components-core";
+import "./theme";