+++ packages/react-types/index.d.ts
@@ -54,18 +54,18 @@ export type Difference<A, B> = Omit<A, keyof B>;
 /**
  * Makes all properties, included nested ones, partial
  */
-type RecursivePartial<T> = {
+type DeepPartial<T> = {
     [P in keyof T]?: T[P] extends (infer U)[]
-        ? RecursivePartial<U>[]
+        ? DeepPartial<U>[]
         : T[P] extends object
-        ? RecursivePartial<T[P]>
+        ? DeepPartial<T[P]>
         : T[P];
 };
 
 /**
  * Makes all properties, included nested ones, partial. Except those of type K
  */
-type RecursivePartialExcept<T, K extends keyof T> = RecursivePartial<T> & Pick<T, K>;
+type DeepPartialExcept<T, K extends keyof T> = RecursivePartial<T> & Pick<T, K>;
 
 /**
  * Makes properties of type K required
@@ -76,6 +76,11 @@ type Demand<T, K extends keyof T> = T & { [P in K]-?: T[P] };
  * Makes properties of type K optional
  */
 type Loosen<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>;
+
+/**
+ * Makes properties of type K optional deeply
+ */
+type LoosenDeeply<T, K extends keyof T> = Omit<T, K> & DeepPartial<Pick<T, K>>;
 
 /**
  * Gets all keys with type undefined of T