+++ packages/react-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Versions
 
+## 1.2.4
+
+[LoosenDeeply]
+[RecursivePartial -> DeepPartial]
+[RecursivePartialExcept -> DeepPartialExcept]
+
 ## 1.2.3
 
 [RecursivePartial]